+++ Runtime/Baker/GT7/GT7Config.cs
@@ -47,7 +47,7 @@ namespace CustomToneMapping.Baker.GT7
                 h = HashUtil.Hash32(h, BlendRatio);
                 h = HashUtil.Hash32(h, FadeStart);
                 h = HashUtil.Hash32(h, FadeEnd);
-                h = HashUtil.Hash32(h, (uint)LutSize);
+                h = HashUtil.Hash32(h, LutSize);
                 return h;
             }
         }
